[FEAT] #8 - collapsing toolbar의 이슈 수정을 위해 상단 메뉴 정보도 recyclerview로 다시 추가 - ToDo: RecyclerView의 스크롤을 통해 상단 Toolbar show/hide 작업 필요
