+++ .idea/deploymentTargetDropDown.xml
@@ -6,12 +6,12 @@
         <type value="RUNNING_DEVICE_TARGET" />
         <deviceKey>
           <Key>
-            <type value="SERIAL_NUMBER" />
-            <value value="adb-R54N905DHRL-RKO4w0._adb-tls-connect._tcp." />
+            <type value="VIRTUAL_DEVICE_PATH" />
+            <value value="$USER_HOME$/.android/avd/Nexus_5X_API_32.avd" />
           </Key>
         </deviceKey>
       </Target>
     </runningDeviceTargetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-08-12T14:22:59.337292Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-08-12T14:53:43.217754Z" />
   </component>
 </project>
