+++ app/src/main/res/layout/item_image_list.xml
@@ -14,5 +14,6 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:adjustViewBounds="true"
+        android:padding="8dp"
         android:scaleType="fitCenter" />
 </layout>
